+++ tests-trybuild/tests/riscv-rt/post_init/fail_arg_type.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::post_init]
+fn before_main(_hart_id: String) {}
+
+fn main() {}