+++ tests-trybuild/tests/riscv-rt/post_init/fail_async.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::post_init]
+async fn before_main() {}
+
+fn main() {}