+++ tests-trybuild/tests/riscv-rt/post_init/pass_unsafe.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::post_init]
+unsafe fn before_main(_hart_id: usize) {}
+
+fn main() {}